Reorganize font_set
